requirements.txt
# Core DroidRun framework
droidrun

# Web framework
flask

# LLM integration
llama-index-llms-openai-like
llama-index-core

# Data validation and serialization
pydantic>=2.0.0

# Environment variable management
python-dotenv

# Standard library (no installation needed)
# - asyncio
# - json
# - os
# - datetime
# - typing
# - queue
# - threading
